requirements.txt
django==2.2.5
djangorestframework
google.cloud
nltk
bs4
gensim
model_mommy
ujson
